@@ -1,6 +1,6 @@
 module github.com/shomali11/slacker/v2
 
-go 1.18
+go 1.21
 
 require (
 	github.com/robfig/cron/v3 v3.0.1
