@@ -1,55 +1,52 @@
 package slacker
 
 import (
-	"log"
+	"log/slog"
 	"os"
 )
 
 type Logger interface {
-	Info(args ...interface{})
-	Infof(format string, args ...interface{})
-	Debug(args ...interface{})
-	Debugf(format string, args ...interface{})
-	Error(args ...interface{})
-	Errorf(format string, args ...interface{})
+	Debug(msg string, args ...any)
+	Info(msg string, args ...any)
+	Warn(msg string, args ...any)
+	Error(msg string, args ...any)
 }
 
 type builtinLogger struct {
 	debugMode bool
-	logger    *log.Logger
+	logger    *slog.Logger
 }
 
 func newBuiltinLogger(debugMode bool) *builtinLogger {
+	opts := &slog.HandlerOptions{
+		Level: getLogLevel(debugMode),
+	}
+
 	return &builtinLogger{
 		debugMode: debugMode,
-		logger:    log.New(os.Stdout, "", log.LstdFlags),
+		logger:    slog.New(slog.NewJSONHandler(os.Stdout, opts)),
 	}
 }
 
-func (l *builtinLogger) Info(args ...interface{}) {
-	l.logger.Println(args...)
+func (l *builtinLogger) Info(msg string, args ...any) {
+	l.logger.Info(msg, args...)
 }
 
-func (l *builtinLogger) Infof(format string, args ...interface{}) {
-	l.logger.Printf(format, args...)
+func (l *builtinLogger) Debug(msg string, args ...any) {
+	l.logger.Debug(msg, args...)
 }
 
-func (l *builtinLogger) Debug(args ...interface{}) {
-	if l.debugMode {
-		l.logger.Println(args...)
+func (l *builtinLogger) Warn(msg string, args ...any) {
+	l.logger.Warn(msg, args...)
+}
+
+func (l *builtinLogger) Error(msg string, args ...any) {
+	l.logger.Error(msg, args...)
+}
+
+func getLogLevel(isDebugMode bool) slog.Level {
+	if isDebugMode {
+		return slog.LevelDebug
 	}
-}
-
-func (l *builtinLogger) Debugf(format string, args ...interface{}) {
-	if l.debugMode {
-		l.logger.Printf(format, args...)
-	}
-}
-
-func (l *builtinLogger) Error(args ...interface{}) {
-	l.logger.Println(args...)
-}
-
-func (l *builtinLogger) Errorf(format string, args ...interface{}) {
-	l.logger.Printf(format, args...)
+	return slog.LevelInfo
 }
